@@ -248,14 +251,14 @@
         <pluginManagement>
 
             <plugins>
-                <!-- Compiler plugin enforces Java 1.6 compatibility and activates
+                <!-- Compiler plugin enforces Java 1.8 compatibility and activates
               annotation processors -->
                 <plugin>
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>2.3.1</version>
                     <configuration>
-                        <source>1.6</source>
-                        <target>1.6</target>
+                        <source>1.8</source>
+                        <target>1.8</target>
                     </configuration>
                 </plugin>
                 <plugin>
