@@ -9,11 +9,12 @@ import javax.ejb.Timer;
 import javax.enterprise.event.Event;
 import javax.enterprise.event.Observes;
 import javax.inject.Inject;
+import javax.ws.rs.core.MultivaluedHashMap;
+import javax.ws.rs.core.MultivaluedMap;
 
 import org.jboss.examples.ticketmonster.model.Booking;
 import org.jboss.examples.ticketmonster.rest.BookingService;
 import org.jboss.examples.ticketmonster.util.CircularBuffer;
-import org.jboss.examples.ticketmonster.util.MultivaluedHashMap;
 import org.jboss.examples.ticketmonster.util.qualifier.BotMessage;
 
 /**
@@ -75,7 +76,7 @@ public class BotService {
             stop();
             // Delete 10 bookings at a time
             while(true) {
-                MultivaluedHashMap<String,String> params = new MultivaluedHashMap<String, String>();
+                MultivaluedMap<String,String> params = new MultivaluedHashMap<>();
                 params.add("maxResults", Integer.toString(10));
                 List<Booking> bookings = bookingService.getAll(params);
                 for (Booking booking : bookings) {
