@@ -6,6 +6,7 @@ import static org.junit.Assert.assertNotNull;
 import java.util.List;
 
 import javax.inject.Inject;
+import javax.ws.rs.core.MultivaluedHashMap;
 import javax.ws.rs.core.MultivaluedMap;
 
 import org.jboss.arquillian.container.test.api.Deployment;
@@ -16,7 +17,6 @@ import org.jboss.examples.ticketmonster.rest.EventService;
 import org.jboss.examples.ticketmonster.rest.MediaService;
 import org.jboss.examples.ticketmonster.service.MediaManager;
 import org.jboss.examples.ticketmonster.service.MediaPath;
-import org.jboss.examples.ticketmonster.util.MultivaluedHashMap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Test;
 import org.junit.runner.RunWith;
